backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
keras>=3.0.0
pillow==10.1.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
